+++ Geometry/Trigonometry/Trig_Functions/secant.py
@@ -1,7 +1,7 @@
 import sys as s
 s.path.append("../..")
 from Numerical_Methods.Constants.Pi_Algorithms.Machin_algo import MachinAlgorithm
-from Geometry.Trigonometry.cosine import Cosine 
+from Geometry.Trigonometry.Trig_Functions.cosine import Cosine 
 
 def Sec(radians): 
     sec = 1/Cosine(radians)
@@ -11,7 +11,7 @@ def Sec(radians):
 
 if __name__ == "__main__":
     precision = 10
-    degree = int(input("Please input angle in degrees: "))
+    degree = float(input("Please input angle in degrees: "))
     radians = degree/180 * float(MachinAlgorithm.calculate_pi(precision))
     sec = Sec(radians)
     print(f'The value of cosec is {sec:.2f}')
